+++ src/clj/rems/api/schema.clj
@@ -5,6 +5,15 @@
             [schema.core :as s])
   (:import (org.joda.time DateTime)))
 
+(def UserId s/Str)
+
+(s/defschema User {:userid UserId})
+
+(s/defschema UserWithAttributes
+  {:userid UserId
+   :name (s/maybe s/Str)
+   :email (s/maybe s/Str)})
+
 (s/defschema CatalogueItemLocalizations
   {s/Keyword {;; TODO :id (it's the catalogue item id) and :langcode
               ;; fields are redundant. If we remove them we can reuse
@@ -51,10 +60,13 @@
 
 (s/defschema Entitlement
   {:resource s/Str
+   :user UserWithAttributes
    :application-id s/Int
    :start DateTime
    :end (s/maybe DateTime)
-   :mail s/Str})
+   :mail (rjs/field s/Str
+                    {:deprecate true
+                     :description "DEPRECATED, will disappear"})}) ;; TODO
 
 (s/defschema EnabledCommand
   {:id s/Int
@@ -107,15 +119,6 @@
    (s/optional-key :license/attachment-filename) LocalizedString
    :license/enabled s/Bool
    :license/archived s/Bool})
-
-(def UserId s/Str)
-
-(s/defschema User {:userid UserId})
-
-(s/defschema UserWithAttributes
-  {:userid UserId
-   :name (s/maybe s/Str)
-   :email (s/maybe s/Str)})
 
 (s/defschema Workflow
   {:id s/Int